Moved functions and renamed
appStatusConfig -> appStatusUtil

Moved all date related functions into input util.
